move prepped structure and docking stuff to modeling and docking envs
