+++ asapdiscovery-alchemy/asapdiscovery/alchemy/cli/prep.py
@@ -308,7 +308,7 @@ def run(
     )
     from asapdiscovery.alchemy.schema.prep_workflow import AlchemyPrepWorkflow
     from asapdiscovery.data.readers.molfile import MolFileFactory
-    from asapdiscovery.data.schema.complex import PreppedComplex
+    from asapdiscovery.modeling.schema import PreppedComplex
     from rich import pretty
     from rich.padding import Padding
