+++ asapdiscovery-data/asapdiscovery/data/tests/conftest.py
@@ -1,5 +1,6 @@
 import pytest
-from asapdiscovery.data.schema.complex import Complex, PreppedComplex
+from asapdiscovery.data.schema.complex import Complex
+from asapdiscovery.modeling.schema import PreppedComplex
 from asapdiscovery.data.schema.ligand import Ligand
 from asapdiscovery.data.services.cdd.cdd_api import CDDAPI
 from asapdiscovery.data.services.services_config import CDDSettings
