+++ asapdiscovery-data/asapdiscovery/data/tests/test_complex_schema.py
@@ -1,6 +1,6 @@
 import pytest
 from asapdiscovery.data.backend.openeye import load_openeye_pdb
-from asapdiscovery.data.schema.complex import Complex, PreppedComplex
+from asapdiscovery.data.schema.complex import Complex
 from asapdiscovery.data.testing.test_resources import fetch_test_file
 from pydantic.v1 import ValidationError
 
@@ -117,42 +117,3 @@ def test_complex_json_file_roundtrip(complex_pdb, tmp_path):
     assert c1 == c2
 
 
-def test_prepped_complex_from_complex(complex_pdb):
-    c1 = Complex.from_pdb(
-        complex_pdb,
-        target_kwargs={"target_name": "test"},
-        ligand_kwargs={"compound_name": "test"},
-    )
-    c2 = PreppedComplex.from_complex(c1, prep_kwargs={})
-    du = c2.target.to_oedu()
-    assert du.HasReceptor()
-    assert du.HasLigand()
-    assert c2.target.target_name == "test"
-    assert c2.ligand.compound_name == "test"
-
-
-def test_prepped_complex_from_oedu_file(complex_oedu):
-    c = PreppedComplex.from_oedu_file(
-        complex_oedu,
-        target_kwargs={"target_name": "test", "target_hash": "test hash"},
-        ligand_kwargs={"compound_name": "test"},
-    )
-    assert c.target.target_name == "test"
-    assert c.ligand.compound_name == "test"
-
-
-def test_prepped_complex_hash(complex_pdb):
-    comp = Complex.from_pdb(
-        complex_pdb,
-        target_kwargs={"target_name": "receptor1"},
-        ligand_kwargs={"compound_name": "ligand1"},
-    )
-    pc = PreppedComplex.from_complex(comp)
-    assert (
-        pc.target.target_hash
-        == "843587eb7f589836d67da772b11584da4fa02fba63d6d3f3062e98c177306abb"
-    )
-    assert (
-        pc.hash
-        == "843587eb7f589836d67da772b11584da4fa02fba63d6d3f3062e98c177306abb+JZJCSVMJFIAMQB-DLYUOGNHNA-N"
-    )
