+++ asapdiscovery-docking/asapdiscovery/docking/schema/pose_generation.py
@@ -12,7 +12,7 @@ from asapdiscovery.data.backend.openeye import (
     set_SD_data,
     smiles_to_oemol,
 )
-from asapdiscovery.data.schema.complex import PreppedComplex
+from asapdiscovery.modeling.schema import PreppedComplex
 from asapdiscovery.data.schema.ligand import Ligand
 from pydantic.v1 import BaseModel, Field, PositiveFloat, PositiveInt
 from rdkit import Chem, RDLogger
