+++ asapdiscovery-docking/asapdiscovery/docking/selectors/selector.py
@@ -2,7 +2,8 @@ import abc
 from typing import Literal, Union
 
 import dask
-from asapdiscovery.data.schema.complex import Complex, PreppedComplex
+from asapdiscovery.data.schema.complex import Complex
+from asapdiscovery.modeling.schema import PreppedComplex
 from asapdiscovery.data.schema.ligand import Ligand
 from asapdiscovery.data.schema.pairs import CompoundStructurePair
 from asapdiscovery.data.util.dask_utils import (
