+++ asapdiscovery-docking/asapdiscovery/docking/tests/test_pairs.py
@@ -1,5 +1,6 @@
 import pytest
-from asapdiscovery.data.schema.complex import Complex, PreppedComplex
+from asapdiscovery.data.schema.complex import Complex
+from asapdiscovery.modeling.schema import PreppedComplex
 from asapdiscovery.data.schema.ligand import Ligand
 from asapdiscovery.data.schema.pairs import CompoundStructurePair
 from asapdiscovery.data.testing.test_resources import fetch_test_file
