+++ asapdiscovery-workflows/asapdiscovery/workflows/tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 from asapdiscovery.data.backend.openeye import oechem
-from asapdiscovery.data.schema.complex import PreppedComplex
+from asapdiscovery.modeling.schema import PreppedComplex
 from asapdiscovery.data.schema.ligand import Ligand
 from asapdiscovery.data.testing.test_resources import fetch_test_file
 from asapdiscovery.docking.docking import DockingInputMultiStructure, DockingInputPair
